rename src/messages to src/messaging
