+++ src/messaging/brandConnection.ts
