+++ src/messaging/serializeConnection.ts
