+++ src/messaging/tests/brandConnection.test.ts
