+++ src/messaging/tests/serializeConnection.test.ts
