+++ src/messaging/workerConnection.ts
